Broke what I had written in the admin anhd models files in the existing app.
Have moved to calling the base app 'core' and will rebuild the model and admin page based on the obvious requirements of the newly staged templates folder.

Templates folder contains the beginnings of the base website. Need to djangofy it by adding template tags and stuff. Make sure the html has the static links in it.